Packages: Move case and mapping methods to js-utils
