+++ packages/js-utils/src/index.ts
@@ -1,2 +1,5 @@
+export { default as camelToSnakeCase } from './camel-to-snake-case';
+export { default as mapRecordKeysRecursively } from './map-record-keys-recursively';
 export { default as shuffle } from './shuffle';
+export { default as snakeToCamelCase } from './snake-to-camel-case';
 export { default as uniqueBy } from './unique-by';
