+++ packages/js-utils/src/snake-to-camel-case.ts
@@ -3,7 +3,7 @@
  *
  * This is designed to work nearly identically to the lodash `camelCase` function.
  */
-export function snakeToCamelCase( snakeCaseString: string ): string {
+export default function snakeToCamelCase( snakeCaseString: string ): string {
 	return snakeCaseString
 		.toLowerCase()
 		.replace( /([-_][a-z0-9])/g, ( group ) =>
