+++ packages/wpcom-checkout/src/index.ts
@@ -26,6 +26,3 @@ export * from './get-introductory-offer-interval-display';
 export * from './join-classes';
 export * from './checkout-line-items';
 export * from './get-country-postal-code-support';
-export * from './map-record-keys-recursively';
-export * from './camel-to-snake-case';
-export * from './snake-to-camel-case';
